requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
# backports.zoneinfo==0.2.1
boto3==1.34.45
botocore==1.34.45
dj-database-url==2.1.0
Django==4.2.9
django-filter==23.5
django-graphql-jwt==0.4.0
django-storages==1.14.2
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.2
pillow==10.2.0
promise==2.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.9.0
urllib3==1.26.18
